Fot rid of all warnings for MSVC
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@39054 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/oligofar/cguidefile.hpp
@@ -43,7 +43,7 @@ public:
 
 	CGuideFile( const string& fileName, CFilter& filter, CSeqIds& seqIds );
     
-    bool NextHit( int ordinal, CQuery * query );
+    bool NextHit( Uint8 ordinal, CQuery * query );
     
 	int GetMaxDist() const { return m_filter->GetMaxDist(); }
 	int GetMinDist() const { return m_filter->GetMinDist(); }
